git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5472 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/MANYBODY/fix_qeq_comb.h
@@ -22,6 +22,7 @@ FixStyle(qeq/comb,FixQEQComb)
 
 #include "stdio.h"
 #include "fix.h"
+#include "lmptype.h"
 
 namespace LAMMPS_NS {
 
@@ -40,7 +41,7 @@ class FixQEQComb : public Fix {
   int me,firstflag;
   double precision;
   int nlevels_respa;
-  double ngroup;
+  bigint ngroup;
   FILE *fp;
 
   class PairComb *comb;
